requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
qdrant-client==1.7.0
python-dotenv==1.0.1
pydantic==1.10.14
python-multipart==0.0.9
pypdf==4.1.0
openai==2.7.1
requests==2.31.0
